+++ mixing-valve/mixing-valve.js
@@ -37,8 +37,8 @@ module.exports = function (RED)
         // ##################
         // # Dynamic config #
         // ##################
-        let time_total = config.time_total;
-        let time_sampling = config.time_sampling;
+        let time_total_s = config.time_total;
+        let time_sampling_s = config.time_sampling;
 
 
         // ##################
@@ -80,13 +80,22 @@ module.exports = function (RED)
             stopSampling();
 
             if (sampling_interval !== null)
+            {
                 clearInterval(sampling_interval);
+                sampling_interval = null;
+            }
 
             if (calibration_timeout !== null)
+            {
                 clearTimeout(calibration_timeout);
+                calibration_timeout = null;
+            }
 
             if (changing_timeout !== null)
+            {
                 clearTimeout(changing_timeout);
+                changing_timeout = null;
+            }
 
             stopChanging();
         });
@@ -197,7 +206,7 @@ module.exports = function (RED)
                 return;
 
             // start sampling
-            sampling_interval = setInterval(sample, time_sampling * 1000);
+            sampling_interval = setInterval(sample, time_sampling_s * 1000);
 
             setStatus();
         }
@@ -223,7 +232,14 @@ module.exports = function (RED)
         {
             // No current temperature available or in calibration => no action
             if (current_temperature === null || calibration_timeout !== null || !node_settings.enabled)
+            {
+                helper.log("No sample possible", {
+                    current_temperature,
+                    calibration_timeout,
+                    enabled: node_settings.enabled
+                });
                 return;
+            }
 
             // +/- 1°C => already good enough, do nothing
             let temp_diff = Math.abs(current_temperature - node_settings.setpoint);
@@ -232,7 +248,7 @@ module.exports = function (RED)
 
             // Calculate change time
             // Change time in ms for 1%
-            let moving_time = time_total * 1000 / 100;
+            let moving_time = time_total_s * 1000 / 100;
             // 0 °C diff => 0% change
             // 20 °C diff => 2% change
             moving_time *= helper.scale(Math.min(temp_diff, 20), 0, 20, 0, 2);
@@ -256,13 +272,14 @@ module.exports = function (RED)
 
         let startChanging = (adjustAction, time_ms) =>
         {
+            helper.log("Start changing", adjustAction, time_ms)
             stopChanging();
 
             // Already oppened/closed
             if (adjustAction == ADJUST_OPEN && node_settings.last_position == 100)
-                time_ms = time_total * 1000 / 200; // Change at least 1/200 => 0.5 %
+                time_ms = time_total_s * 1000 / 200; // Change at least 1/200 => 0.5 %
             else if (adjustAction == ADJUST_CLOSE && node_settings.last_position == 0)
-                time_ms = time_total * 1000 / 200; // Change at least 1/200 => 0.5 %
+                time_ms = time_total_s * 1000 / 200; // Change at least 1/200 => 0.5 %
 
             adjusting_start_time = Date.now();
             if (adjustAction == ADJUST_OPEN)
@@ -274,9 +291,12 @@ module.exports = function (RED)
             changing_timeout = setTimeout(() =>
             {
                 changing_timeout = null;
+                adjusting = null;
                 stopChanging();
                 setStatus();
             }, time_ms);
+
+            setStatus();
         }
 
         let stopChanging = () =>
@@ -296,7 +316,7 @@ module.exports = function (RED)
             let time_passed = (Date.now() - adjusting_start_time) / 1000;
             adjusting_start_time = null;
 
-            let changed_value = (time_passed / time_total) * 100; // calculate in % value (0-100)
+            let changed_value = (time_passed / time_total_s) * 100; // calculate in % value (0-100)
             if (adjusting == ADJUST_OPEN)
                 node_settings.last_position += changed_value;
             else
@@ -305,7 +325,7 @@ module.exports = function (RED)
             // Only values from 0 to 100 are allowed
             node_settings.last_position = Math.min(Math.max(node_settings.last_position, 0), 100);
 
-            node.send([{ payload: false }, { payload: false }, { payload: node_settings.last_position.toFixed(1) }]);
+            node.send([{ payload: false }, { payload: false }, { payload: helper.toFixed(node_settings.last_position, 1) }]);
 
             setStatus();
         }
@@ -331,7 +351,7 @@ module.exports = function (RED)
                     stopSampling();
 
                 setStatus();
-            }, time_total * 1000);
+            }, time_total_s * 1000);
         }
 
         let doOffMode = () =>
@@ -339,11 +359,11 @@ module.exports = function (RED)
             switch (node_settings.off_mode)
             {
                 case "OPEN":
-                    startChanging(ADJUST_OPEN, time_total * 1000);
+                    startChanging(ADJUST_OPEN, time_total_s * 1000);
                     break;
 
                 case "CLOSE":
-                    startChanging(ADJUST_CLOSE, time_total * 1000);
+                    startChanging(ADJUST_CLOSE, time_total_s * 1000);
                     break;
 
                 case "NOTHING":
@@ -357,9 +377,9 @@ module.exports = function (RED)
             if (calibration_timeout !== null)
                 node.status({ fill: "yellow", shape: "ring", text: helper.getCurrentTimeForStatus() + ": In calibration" });
             else if (changing_timeout != null)
-                node.status({ fill: node_settings.enabled ? "green" : "red", shape: "ring", text: helper.getCurrentTimeForStatus() + ": " + (node_settings.valve_mode == "HEATING" ? "🔥" : "❄️") + "  " + (adjusting == ADJUST_OPEN ? "Opening" : "Closing") + ", Set: " + node_settings.setpoint?.toFixed(1) + "°C, Cur: " + current_temperature?.toFixed(1) + "°C, Pos: " + node_settings.last_position?.toFixed(1) + "%" });
+                node.status({ fill: node_settings.enabled ? "green" : "red", shape: "ring", text: helper.getCurrentTimeForStatus() + ": " + (node_settings.valve_mode == "HEATING" ? "🔥" : "❄️") + "  " + (adjusting == ADJUST_OPEN ? "Opening" : "Closing") + ", Set: " + helper.toFixed(node_settings.setpoint, 1) + "°C, Cur: " + helper.toFixed(current_temperature, 1) + "°C, Pos: " + helper.toFixed(node_settings.last_position, 1) + "%" });
             else
-                node.status({ fill: node_settings.enabled ? "green" : "red", shape: "dot", text: helper.getCurrentTimeForStatus() + ": " + (node_settings.valve_mode == "HEATING" ? "🔥" : "❄️") + " Set: " + node_settings.setpoint?.toFixed(1) + "°C, Cur: " + current_temperature?.toFixed(1) + "°C, Pos: " + node_settings.last_position?.toFixed(1) + "%" });
+                node.status({ fill: node_settings.enabled ? "green" : "red", shape: "dot", text: helper.getCurrentTimeForStatus() + ": " + (node_settings.valve_mode == "HEATING" ? "🔥" : "❄️") + " Set: " + helper.toFixed(node_settings.setpoint, 1) + "°C, Cur: " + helper.toFixed(current_temperature, 1) + "°C, Pos: " + helper.toFixed(node_settings.last_position, 1) + "%" });
         }
 
         if (node_settings.last_position == null)
@@ -370,7 +390,7 @@ module.exports = function (RED)
         else if (node_settings.enabled)
         {
             startSampling();
-            node.send([null, null, { payload: node_settings.last_position.toFixed(1) }]);
+            node.send([null, null, { payload: helper.toFixed(node_settings.last_position, 1) }]);
         }
 
         setStatus();